+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version History
 v5.30.1
 -------
 
+* Fix CameraCableWrap UI swapped limits and floating points `<https://github.com/lsst-ts/LOVE-frontend/pull/628>`_
 * Fix value of MTHexapod_logevent_compensatedPosition.w setting `<https://github.com/lsst-ts/LOVE-frontend/pull/626>`_
 
 v5.30.0